+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Setting: `[qol] fast_countdown_enable`
 - Setting: `[qol] fast_countdown_duration`
 
+### Changed
+
+- Improved framerate limiter pacing and CPU impact
+
 ## [0.1.2] -- 2024-04-23
 
 ### Added