+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.3] -- 2024-04-30
 
 ### Added
 
@@ -66,7 +66,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Triggered race events displayed on UI
 - Rainbow-colored race UI elements
 
-[unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/0.1.2...HEAD
+[unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/0.1.3...HEAD
+[0.1.3]: https://github.com/olivierlacan/keep-a-changelog/compare/0.1.2...0.1.3
 [0.1.2]: https://github.com/olivierlacan/keep-a-changelog/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/olivierlacan/keep-a-changelog/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/olivierlacan/keep-a-changelog/releases/tag/0.1.0